RelevanceVectorMachine.jl Documentation

RelevanceVectorMachine.jl Documentation

Quickstart

Functions

References

Quickstart

See the examples folder. If you clone the repository, then you should be able to run the examples using Pluto.jl notebooks.

Functions

RelevanceVectorMachine.RVM — Type

RVM

An instance of a relevance vector machine. Can be instantiated using rvm.

RelevanceVectorMachine.rvm — Function

rvm(formula::FormulaTerm, data, mode = "regression", max_iters = 1000)

Initialize and train a relevance vector machine, using the variables specified in formula with the data provided in data. mode can either be "regression" or "classification".

RelevanceVectorMachine.predict — Method

predict(rvm::RVM, X)

Given a table X and relevance vector machine rvm, compute predictions for X.

RelevanceVectorMachine.posterior — Method

posterior(rvm::RVM)

Returns a normal distribution with mean μ and covariance Σ corresponding to the distribution of the weight parameter (a vector) of the relevance vector machine.

References

Tipping, M.E. &amp; Faul, A.C.. (2003). Fast Marginal Likelihood Maximisation for Sparse Bayesian Models. Proceedings of the Ninth International Workshop on Artificial Intelligence and Statistics, in Proceedings of Machine Learning Research R4:276-283 Available from https://proceedings.mlr.press/r4/tipping03a.html. Reissued by PMLR on 01 April 2021.